pyproject.toml
[project]
name = "beikecrawlermap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.5",
    "fastapi>=0.116.1",
    "pandas>=2.3.1",
    "pydoll-python>=2.6.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.48.1",
    "uvicorn>=0.35.0",
]
